requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
Pillow==12.0.0
keras==3.11.3
matplotlib==3.10.7
streamlit==1.50.0
tensorflow==2.20.0
